Extract logging setup into utils/web_logging.py
Continues the main.py split. Moves the colored console formatter, the
WebLogHandler + in-memory log ring buffer, get_recent_logs, the lazy
handler attach, and the uvicorn log config into utils/web_logging.py.
The version-stamped format string is now built by configure_logging(),
which main.py calls at startup. Drops the now-unused itertools/deque
imports from main.py. Behaviour unchanged.
